extracted data from view
