+++ HealthTracker/healthtracker.client/src/data/models/friendModel.ts
